On branch restapi №13 Изменения: - Написан EmployeesService в api/services/employees.py; - Написан интерфейс DataConverter для конвертации данных из DTO в модели или словари и наоборот в api/schemas/commons.py, там же написан AbstractDataConverter, возможно, в будущем DataConverter можно будет прокидывать в экземпляр uow или в сервис; - Написан менеджер репозиториев RepoManager в api/utils/repositories_manager.py, он будет использоваться для добавления репозиториев экземпляру класса UnitOfWork; - Написан абстрактный класс uow и класс uow для SqlAlchemy в api/utils/unitofwork.py; - Исправлены опечатки в SqlAlchemy моделях в api/models/*.py, но, возможно, не все; - Изменён api/routers/profile.py, теперь приложение FastAPI можно запустить и протестировать, но пока что протестировать можно только один эндпоинт;

+++ api/database.py
@@ -11,7 +11,7 @@ engine = create_engine(
 
 session_factory = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-idpk = Annotated[BigInteger, mapped_column(primary_key=True)]
+idpk = Annotated[int, mapped_column(primary_key=True)]
 created_at = Annotated[TIMESTAMP, mapped_column(TIMESTAMP, server_default=text('TIMEZONE("utc+5", now()'))]
